Add support for retriggering dependent workflows
Currently when using an ExternalTaskSensor sensor, to have a dag wait
for the completion of a task in another dag, there is no way, when
clearing the task that is depended on, to also clear the ExternalTaskSensor
task (and its downstream/upstream tasks).

However, that might be important when teams have separate responsabilities
and create multi-staged data pipelines. Let's say there is a team reponsible
for transforming logs to make it available in hadoop land. They maintain an
hourly dag processing the raw logs. Other teams wait on the completion of
this dag using an ExternalTaskSensor. If the logging team realizes that
some logs where corrupted, not only do they have to rerun the task that
processed these logs but also all the dags that use these logs to generate
higher level data.

This commit tries to solve this problem by introducing the notion of dag
descencents. The descendents of dag A are the dags that have an
ExternalTaskSensor pointing to a task of dag A.
Then, when clearing a task in the view, an option is added to also clear
the tasks of the descendents (having the upstream, downstream, recursive,
future and past flags act accordingly).