BashOperator: Added process_output parameter to the BashOperator
xcom push for a bash operator uses the last line of the process output
as the xcom value. This can be limiting, with the only good solution
being to wrap any program that we do not directly control with a custom
script that is responsible for running the target program and processing
its output to produce a single line, which in turn will be used for xcom
push.

This patch inverts the approach by allowing the programmer to optionally
specify a Python function to the the BashOperator task which can then do
the same thing as the wrapper script would -- post process the output of
the target program to be used as the value of the xcom push. The
advantage here is that this is pure Python (hence the full language is
at one's disposal for processing the bash process output), more
readable (because the post-processing code lives alongside the DAG
definition code), and more amenable to code reuse.